upgraded signup page and folder structure
