--- client/src/Auth/firebase.config.ts
@@ -1,18 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyBu47oy8N5BcjPEn2_wAPNs9oHVpvDY08A",
-  authDomain: "hospital-management-4e9fc.firebaseapp.com",
-  projectId: "hospital-management-4e9fc",
-  storageBucket: "hospital-management-4e9fc.appspot.com",
-  messagingSenderId: "157172134842",
-  appId: "1:157172134842:web:290fa728a01c69f215e383"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-export default app
