+++ client/src/main.tsx
@@ -5,7 +5,8 @@ import "./index.css";
 // router setup
 import { RouterProvider } from "react-router-dom";
 import router from "./routes/Routes.tsx";
-import AuthProvider from "./Auth/AuthProvider.tsx";
+import AuthProvider from "./providers/AuthProvider.tsx";
+
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
